api/requirements.txt
fastapi==0.78.0
starlette==0.19.1
pydantic<2
python-dotenv
langchain==0.2.5
langchain-openai==0.1.14
langchain-core==0.2.8
langchain-text-splitters==0.2.1
openai>=1.32.0,<2
anyio<4
